Refactor saveJSON function
