+++ src/AppModel.js
@@ -14,6 +14,7 @@ import {
   moveItem,
   updateRestrictedDropId,
 } from './redux';
+import saveJSON from './saveJSON';
 
 const AppModel = () => {
   const items = useSelector((state) => state.block.items);
@@ -95,29 +96,6 @@ const AppModel = () => {
       source.index,
     ));
   };
-
-  function saveJSON(data, filename) {
-    if (!data) {
-      console.error('No data');
-      return;
-    }
-
-    if (!filename) { filename = 'console.json'; }
-
-    if (typeof data === 'object') {
-      data = JSON.stringify(data, undefined, 4);
-    }
-
-    const blob = new Blob([data], { type: 'text/json' });
-    const e = document.createEvent('MouseEvents');
-    const a = document.createElement('a');
-
-    a.download = filename;
-    a.href = window.URL.createObjectURL(blob);
-    a.dataset.downloadurl = ['text/json', a.download, a.href].join(':');
-    e.initMouseEvent('click', true, false, window, 0, 0, 0, 0, 0, false, false, false, false, 0, null);
-    a.dispatchEvent(e);
-  }
 
   return (
     <div className="App">
